Площадь треугольника.

+++ Test_03-01/Test_03-01.py
@@ -1,6 +1,6 @@
-a = int(input())
-b = int(input())
-c = int(input())
+a = float(input())
+b = float(input())
+c = float(input())
 p = (a + b + c) / 2
 s = (p * (p - a) * (p - b) * (p - c)) ** (1/2)
 print('{:.6f}'.format(s))
